requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
Flask-Migrate
psycopg2-binary
redis
celery
authlib
python-dotenv
gunicorn
google-cloud-storage
dnspython
pillow-heif
extract-msg
pandas
openpyxl
requests
flask-wtf
flask-cors
flask-bootstrap
email_validator
gevent
pdfplumber
pdf2image
Pillow
pytesseract
squareup
